added some tests 4

+++ src/test/java/hexlet/code/app/util/ModelGenerator.java
@@ -36,7 +36,7 @@ public class ModelGenerator {
                 .ignore(Select.field(Task::getTaskStatus))
                 .ignore(Select.field(Task::getLabels))
                 .supply(Select.field(Task::getName), () -> faker.lorem().word())
-                .supply(Select.field(Task::getIndex), () -> faker.number().randomDigitNotZero())
+                .supply(Select.field(Task::getIndex), () -> (long) faker.number().randomDigitNotZero())
                 .supply(Select.field(Task::getDescription), () -> faker.lorem().sentence())
                 .toModel();
